project was restructured
