--- PolygonsParallelToLine/resources.py
@@ -1,88 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Resource object code
-#
-# Created: Вс 3. апр 19:25:45 2016
-#      by: The Resource Compiler for PyQt (Qt v4.8.5)
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt4 import QtCore
-
-qt_resource_data = "\
-\x00\x00\x02\x9c\
-\x89\
-\x50\x4e\x47\x0d\x0a\x1a\x0a\x00\x00\x00\x0d\x49\x48\x44\x52\x00\
-\x00\x00\x20\x00\x00\x00\x20\x08\x06\x00\x00\x00\x73\x7a\x7a\xf4\
-\x00\x00\x00\x06\x62\x4b\x47\x44\x00\x00\x00\x00\x00\x00\xf9\x43\
-\xbb\x7f\x00\x00\x00\x09\x70\x48\x59\x73\x00\x00\x12\x65\x00\x00\
-\x12\x65\x01\x45\x35\xed\x56\x00\x00\x00\x07\x74\x49\x4d\x45\x07\
-\xe0\x03\x1f\x07\x15\x35\xf1\x03\xb8\x89\x00\x00\x02\x29\x49\x44\
-\x41\x54\x58\xc3\xed\xd7\x3f\x6c\x4e\x61\x14\xc7\xf1\xcf\xad\x52\
-\x43\x35\x78\x1b\xc5\xd0\x4a\x48\x10\x83\x89\x16\x9d\xda\x88\x08\
-\x89\x46\x74\x21\x2c\x16\x83\x41\x2c\x22\x69\x9a\x30\x48\x18\x24\
-\x24\xd2\xc4\x44\x2c\x12\x4d\x3b\xb0\x50\x69\x10\xda\xa9\x04\x91\
-\x10\x4b\x07\x4a\x48\x0d\x16\x09\xc3\x31\xbc\xb7\xf2\xf6\xdf\xeb\
-\xbe\xed\xdb\x76\xe0\x49\xee\x72\xcf\xc9\xef\x7b\xce\x79\x9e\x7b\
-\xcf\x79\xf8\xd7\x57\x32\x07\x9a\x2b\x71\x10\x3b\xb0\x16\x81\x11\
-\x3c\x47\x37\xbe\x14\x3a\x2f\x2a\x23\x78\x29\xce\xe1\x16\xaa\xf0\
-\x0c\xf7\xf1\x14\xdf\xd0\x8a\x2b\x58\x92\xda\xa2\x9c\x59\xd7\x60\
-\x30\xcd\x70\x75\x11\xbf\x3a\x3c\xc2\xb5\x59\x13\x83\xe6\xc8\x0b\
-\xc2\x26\x74\x64\xdc\xd2\x6a\xbc\xc7\xae\xd9\xc0\x9b\x82\xaf\xc1\
-\xbb\xa0\x61\x06\x12\xa7\xd0\x35\x53\xf8\xb6\xe0\x63\xd0\x18\x1c\
-\x0b\x3e\x04\x5b\x4b\x94\x69\xc5\xc3\x99\xc0\xb7\xa4\xc0\xd6\x82\
-\x77\x6d\xc1\xa7\x60\x67\x09\x52\x6d\xe8\x2d\x15\xbe\x31\xcd\xfc\
-\xc0\x14\xb6\x96\xe0\x73\xb0\x27\xa3\x5c\x27\x2e\x96\x02\xaf\x0f\
-\x86\x83\xa3\x45\x7c\xb6\xa7\x95\x68\xcf\x20\xf9\x1a\x4d\x59\xe1\
-\x2b\x82\x97\xc1\x99\x0c\xbe\x8d\x69\x25\x0e\x17\x71\xdb\x8f\x57\
-\x99\xbe\x98\xa0\x3a\x18\x0c\x2e\x94\x50\xad\xae\x60\x28\xa8\x9c\
-\x6c\xb2\x18\x6f\xb0\x2f\x2b\x7c\x20\xb8\x54\x02\xbc\x23\x78\x11\
-\x2c\x9f\xda\x2c\x70\x37\x8b\x50\x55\xd0\x17\x5c\x8f\x8c\xfd\x22\
-\x38\x1d\xbc\x0d\x56\x4d\xef\xf2\xe7\x29\x2a\x54\x19\xf4\x06\xb7\
-\x23\x63\xaf\x08\x8e\xa7\x87\xb4\xbe\xb8\xdb\xe4\x20\x92\x09\x1e\
-\x15\xb8\x81\x1c\xda\x6a\xfb\x73\x3f\xcb\xdd\x2a\x47\x5b\x46\xc7\
-\x31\x2b\x26\xd8\xaf\xca\xff\x5a\x0f\x25\xfc\x9a\x8f\x79\xa0\xb2\
-\x20\xfb\xf3\xe9\xb7\xd9\x92\xf0\xa3\x58\xd4\x63\xcd\x08\x3d\x68\
-\x4b\x18\xf8\x1b\x28\xd7\x9f\x8b\x69\x03\x08\x4e\xe2\x08\x9a\x13\
-\xbe\x67\xd8\xf3\xcd\xb8\x83\x13\x59\xe0\xc5\xd6\xd8\x16\x8c\xa4\
-\x6d\x72\x7d\x06\xf8\xba\x74\xd0\x38\x9b\xe4\x2b\x60\xd6\x01\xa4\
-\x42\xed\xe8\x89\xfc\x38\x35\x1d\xbc\x01\x8f\xd1\x99\xe4\x0f\x6b\
-\xf9\xce\x40\xc2\x93\x60\x2f\xee\x05\xcb\x12\x6e\x4e\x80\xd7\xa1\
-\x0f\x97\xcb\x05\x1f\x17\x40\x1a\xc4\x50\xd0\x82\x07\x41\x4d\x6d\
-\x41\x3f\x48\xcb\xde\x9d\xe4\xe7\xba\xb9\x9d\x8a\xd3\x7d\xee\xab\
-\xed\xcf\x6d\x98\xef\xff\xc0\x58\x54\xc3\xd8\xbd\xe0\x97\x86\x60\
-\x4d\x70\x35\xf2\xa3\xf4\x9c\x5e\x4c\x62\x81\x72\x4c\x2a\x16\xba\
-\xca\x0b\x1e\xc0\xff\xf5\x1b\xf0\xaa\xb7\x3c\x5d\xe5\x6e\x84\x00\
-\x00\x00\x00\x49\x45\x4e\x44\xae\x42\x60\x82\
-"
-
-qt_resource_name = "\
-\x00\x07\
-\x07\x3b\xe0\xb3\
-\x00\x70\
-\x00\x6c\x00\x75\x00\x67\x00\x69\x00\x6e\x00\x73\
-\x00\x16\
-\x0f\x62\x9d\x05\
-\x00\x50\
-\x00\x6f\x00\x6c\x00\x79\x00\x67\x00\x6f\x00\x6e\x00\x73\x00\x50\x00\x61\x00\x72\x00\x61\x00\x6c\x00\x6c\x00\x65\x00\x6c\x00\x54\
-\x00\x6f\x00\x4c\x00\x69\x00\x6e\x00\x65\
-\x00\x08\
-\x0a\x61\x5a\xa7\
-\x00\x69\
-\x00\x63\x00\x6f\x00\x6e\x00\x2e\x00\x70\x00\x6e\x00\x67\
-"
-
-qt_resource_struct = "\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
-\x00\x00\x00\x14\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
-\x00\x00\x00\x46\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-"
-
-def qInitResources():
-    QtCore.qRegisterResourceData(0x01, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-def qCleanupResources():
-    QtCore.qUnregisterResourceData(0x01, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-qInitResources()
