+++ polygons_parallel_to_line_dialog.py
