+++ polygons_parallel_to_line_dialog_base.ui
